Preparation of code to receive generated code
